add a few tests

+++ ElectronicObserver/Utility/Data/ExerciseExp.cs
@@ -1,12 +1,15 @@
 ﻿namespace ElectronicObserver.Utility.Data;
 
+/// <summary>
+/// These values aren't floored so that we can apply the flagship modifier in later calculations.
+/// </summary>
 public class ExerciseExp
 {
-	public int BaseA { get; init; }
-	public int BaseS { get; init; }
+	public double BaseA { get; init; }
+	public double BaseS { get; init; }
 
-	public int? TrainingCruiserSurfaceA { get; set; }
-	public int? TrainingCruiserSurfaceS { get; set; }
-	public int? TrainingCruiserSubmarineA { get; set; }
-	public int? TrainingCruiserSubmarineS { get; set; }
+	public double? TrainingCruiserSurfaceA { get; set; }
+	public double? TrainingCruiserSurfaceS { get; set; }
+	public double? TrainingCruiserSubmarineA { get; set; }
+	public double? TrainingCruiserSubmarineS { get; set; }
 }
